Moved listItems to their own component. Displays!
